Added super admin check to admin views
