+++ open_event/views/admin/super_admin/users.py
@@ -1,17 +1,9 @@
-from flask import request, url_for, redirect
 from flask_admin import expose
-from flask.ext import login
+
+from open_event.views.admin.super_admin.super_admin_base import SuperAdminBaseView
 from ....helpers.data_getter import DataGetter
-from flask_admin import BaseView
 
-
-class SuperAdminUsersView(BaseView):
-    def is_accessible(self):
-        return login.current_user.is_authenticated
-
-    def _handle_view(self, name, **kwargs):
-        if not self.is_accessible():
-            return redirect(url_for('admin.login_view', next=request.url))
+class SuperAdminUsersView(SuperAdminBaseView):
 
     @expose('/')
     def index_view(self):
